requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
yfinance==0.2.36
fredapi==0.5.1
pandas-datareader==0.10.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0 